Test ts lint error wip fix
Put some stuff back

Put some more stuff back
